data-harvester/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.1
numpy==1.26.0
aiohttp==3.8.6
httpx==0.25.0
lxml==4.9.3
pytest==7.4.2
pytest-asyncio==0.21.1 